+++ util-jackson/src/test/scala/com/twitter/util/jackson/JSONTest.scala
@@ -1,29 +1,26 @@
 package com.twitter.util.jackson
 
+import com.twitter.io.Buf
+import java.io.{ByteArrayInputStream, File => JFile}
 import org.junit.runner.RunWith
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatestplus.junit.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
-class JSONTest extends AnyFunSuite with Matchers {
+class JSONTest extends AnyFunSuite with Matchers with FileResources {
 
   test("JSON#parse 1") {
     JSON.parse[Map[String, Int]]("""{"a": 1, "b": 2}""") match {
       case Some(map) =>
-        map("a") should equal(1)
-        map("b") should equal(2)
+        map should be(Map("a" -> 1, "b" -> 2))
       case _ => fail()
     }
   }
 
   test("JSON#parse 2") {
     JSON.parse[Seq[String]]("""["a", "b", "c"]""") match {
-      case Some(seq) =>
-        seq.size should equal(3)
-        seq.head should be("a")
-        seq(1) should be("b")
-        seq.last should be("c")
+      case Some(Seq("a", "b", "c")) => // pass
       case _ => fail()
     }
   }
@@ -43,6 +40,40 @@ class JSONTest extends AnyFunSuite with Matchers {
         clazz.name.isEmpty should be(true)
         clazz.make should equal(CarMakeEnum.vw)
       case _ => fail()
+    }
+  }
+
+  test("JSON#parse 5") {
+    val buf = Buf.Utf8("""{"id": "abcd1234"}""")
+    JSON.parse[FooClass](buf) match {
+      case Some(foo) =>
+        foo.id should equal("abcd1234")
+      case _ => fail()
+    }
+  }
+
+  test("JSON#parse 6") {
+    val inputStream = new ByteArrayInputStream("""{"id": "abcd1234"}""".getBytes("UTF-8"))
+    try {
+      JSON.parse[FooClass](inputStream) match {
+        case Some(foo) =>
+          foo.id should equal("abcd1234")
+        case _ => fail()
+      }
+    } finally {
+      inputStream.close()
+    }
+  }
+
+  test("JSON#parse 7") {
+    withTempFolder {
+      val file: JFile =
+        writeStringToFile(folderName, "test-file", ".json", """{"id": "999999999"}""")
+      JSON.parse[FooClass](file) match {
+        case Some(foo) =>
+          foo.id should equal("999999999")
+        case _ => fail()
+      }
     }
   }
 
@@ -77,5 +108,13 @@ class JSONTest extends AnyFunSuite with Matchers {
     JSON.prettyPrint(FooClass("abcd1234")) should equal("""{
         |  "id" : "abcd1234"
         |}""".stripMargin)
+  }
+
+  test("JSON.Resource#parse resource") {
+    JSON.Resource.parse[FooClass]("/test.json") match {
+      case Some(foo) =>
+        foo.id should equal("55555555")
+      case _ => fail()
+    }
   }
 }